# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema de Venta de Pasajes — API
  description: >
    Contrato formal de la API REST para el Sistema de Venta de Pasajes
    de la Flota Mariscal Santa Cruz.  Cubre Etapas 1 (Catálogos) y
    2 (Creación de Viaje) según spec.md.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Desarrollo local
...
paths:
...
  /viajes/{id}/boletos:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Viajes]
      summary: Listar boletos vendidos de un viaje (Etapa 5)
      operationId: listarBoletosViaje
      responses:
        '200':
          description: Lista de boletos del viaje
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BoletoViaje'
        '404':
          $ref: '#/components/responses/NotFound'

  # ──────────────────── BOLETOS (Etapa 3) ────────────────────
...
  /despachos:
    post:
      tags: [Despachos]
      summary: Crear despacho y validar boletos
      operationId: crearDespacho
      description: >
        Crea el despacho (estado 'iniciado') y automáticamente inserta todos los
        boletos del viaje en despacho_pasajero con estado_presencia = 'ausente'.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DespachoCreate'
      responses:
        '201':
          description: Despacho creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Despacho'
        '400':
          $ref: '#/components/responses/NotFound'
  /despachos/{id}/pasajeros/{boleto_id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
      - name: boleto_id
        in: path
        required: true
        schema:
          type: integer
    patch:
      tags: [Despachos]
      summary: Actualizar estado de presencia de un pasajero
      operationId: actualizarPresencia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                estado_presencia:
                  type: string
                  enum: [presente, ausente, no_show]
              required: [estado_presencia]
      responses:
        '200':
          description: Presencia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DespachoPasajero'
        '404':
          $ref: '#/components/responses/NotFound'
  /despachos/{id}/finalizar:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    patch:
      tags: [Despachos]
      summary: Finalizar despacho (El bus arranca)
      operationId: finalizarDespacho
      description: >
        Cambia el estado del despacho a 'despachado' y CRÍTICAMENTE
        cambia el estado del viaje asociado a 'en_ruta'.
      responses:
        '200':
          description: Despacho finalizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Despacho'
        '404':
          $ref: '#/components/responses/NotFound'

# ═══════════════════════════════════════════════════════════════════════
#  COMPONENTS
# ═══════════════════════════════════════════════════════════════════════
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Ciudad ──
...
    BoletoViaje:
      type: object
      description: Dato simplificado del boleto para el despachador
      properties:
        id:
          type: integer
        codigo_unico:
          type: string
        asiento_id:
          type: integer
        numero_asiento:
          type: string
        nombre_pasajero:
          type: string
        numero_documento:
          type: string
    Despacho:
      type: object
      properties:
        id:
          type: integer
        viaje_id:
          type: integer
        despachador_id:
          type: integer
        fecha_despacho:
          type: string
        estado:
          type: string
          enum: [iniciado, tripulacion_validada, pasajeros_validados, despachado]
        observaciones:
          type: string
    DespachoCreate:
      type: object
      properties:
        viaje_id:
          type: integer
        despachador_id:
          type: integer
      required: [viaje_id, despachador_id]
    DespachoPasajero:
      type: object
      properties:
        id:
          type: integer
        despacho_id:
          type: integer
        boleto_id:
          type: integer
        estado_presencia:
          type: string
          enum: [presente, ausente, no_show]
        hora_validacion:
          type: string